Add logic for writing to notion
Note this is as-of-yet untested. Still have some work to do in terms of the check before writing to Notion.

Only handles writing new items. If it finds a match, will skip for now.

Also, still need to figure out when to re-write to Notion. Maybe I can add a 'Date Added' and have it be automatically set. Then, if its >### days, overwrite properties instead.